[split/13] move rest of expr to expr_fn in datafusion-expr module (#1794)
* move expr functions to datafusion-expr expr_fn

* add module level comments
